% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{cellsFailQC}
\alias{cellsFailQC}
\alias{cellsFailQC,RNAseq-method}
\title{Access cellsFailQC data}
\usage{
cellsFailQC(object)

\S4method{cellsFailQC}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
cellsFailQC.
}
\description{
Accesses \code{cellsFailQC} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
