% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{cellsFailQC<-}
\alias{cellsFailQC<-}
\alias{cellsFailQC<-,RNAseq,vector-method}
\title{Modify cellsFailQC data}
\usage{
cellsFailQC(object) <- value

\S4method{cellsFailQC}{RNAseq,vector}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new cellsFailQC data for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies cellsFailQC.
}
\description{
Modifies \code{cellsFailQC} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
