% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{CellsSizeFac}
\alias{CellsSizeFac}
\alias{CellsSizeFac,RNAseq-method}
\title{Access CellsSizeFac data}
\usage{
CellsSizeFac(object)

\S4method{CellsSizeFac}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
CellsSizeFac
}
\description{
Accesses \code{CellsSizeFac} slot in \code{RNAseq} class object
}
\author{
Blanca Pijuan Sala.
}
