% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{CellsSizeFac<-}
\alias{CellsSizeFac<-}
\alias{CellsSizeFac<-,RNAseq,list-method}
\title{Modify CellsSizeFac data}
\usage{
CellsSizeFac(object) <- value

\S4method{CellsSizeFac}{RNAseq,list}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new CellsSizeFac for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies CellsSizeFac.
}
\description{
Modifies \code{CellsSizeFac} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
