% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{clustering}
\alias{clustering}
\alias{clustering,RNAseq-method}
\title{Access clustering data}
\usage{
clustering(object)

\S4method{clustering}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
clustering.
}
\description{
Accesses \code{clustering} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
