% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{clustering<-}
\alias{clustering<-}
\alias{clustering<-,RNAseq-method}
\title{Modify clustering data}
\usage{
clustering(object) <- value

\S4method{clustering}{RNAseq}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new clustering data for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies clustering.
}
\description{
Modifies \code{clustering} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
