% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{CountsFeat}
\alias{CountsFeat}
\alias{CountsFeat,RNAseq-method}
\title{Access CountsFeat data}
\usage{
CountsFeat(object)

\S4method{CountsFeat}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
CountsFeat.
}
\description{
Accesses \code{CountsFeat} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
