% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{CountsFeat<-}
\alias{CountsFeat<-}
\alias{CountsFeat<-,RNAseq,matrix-method}
\title{Modify CountsFeat data}
\usage{
CountsFeat(object) <- value

\S4method{CountsFeat}{RNAseq,matrix}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Matrix or dataframe containing new CountsFeat data
for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies CountsFeat.
}
\description{
Modifies \code{CountsFeat} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
