% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meta_countsMatrix.R
\docType{data}
\name{countsMatrix}
\alias{countsMatrix}
\title{Example of counts}
\source{
Blanca Pijuan-Sala (Gottgens Lab)
}
\usage{
data(countsMatrix)
}
\description{
Example of counts
}
\author{
Blanca Pijuan-Sala
}
