% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{countsNorm}
\alias{countsNorm}
\alias{countsNorm,RNAseq-method}
\title{Access countsNorm data}
\usage{
countsNorm(object)

\S4method{countsNorm}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
countsNorm.
}
\description{
Accesses \code{countsNorm} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
