% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{countsNorm<-}
\alias{countsNorm<-}
\alias{countsNorm<-,RNAseq,matrix-method}
\title{Modify countsNorm data}
\usage{
countsNorm(object) <- value

\S4method{countsNorm}{RNAseq,matrix}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Matrix or dataframe containing new countsNorm data
for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies countsNorm.
}
\description{
Modifies \code{countsNorm} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
