% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{countsRaw}
\alias{countsRaw}
\alias{countsRaw,RNAseq-method}
\title{Access countsRaw data}
\usage{
countsRaw(object)

\S4method{countsRaw}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
countsRaw
}
\description{
Accesses \code{countsRaw} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
