% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{countsRaw<-}
\alias{countsRaw<-}
\alias{countsRaw<-,RNAseq,matrix-method}
\title{Modify countsRaw data}
\usage{
countsRaw(object) <- value

\S4method{countsRaw}{RNAseq,matrix}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Matrix or dataframe containing new countsRaw data for 
\code{RNAseq} object.}
}
\value{
It does not display any value but modifies countsRaw.
}
\description{
Modifies \code{countsRaw} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala
}
