% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{difmap}
\alias{difmap}
\alias{difmap,RNAseq-method}
\title{Access difmap data}
\usage{
difmap(object)

\S4method{difmap}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
difmap.
}
\description{
Accesses \code{difmap} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
