% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{difmap<-}
\alias{difmap<-}
\alias{difmap<-,RNAseq-method}
\title{Modify difmap data}
\usage{
difmap(object) <- value

\S4method{difmap}{RNAseq}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new difmap data for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies difmap.
}
\description{
Modifies \code{difmap} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
