% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_findHVG_DM.R
\name{findHVG.DM}
\alias{findHVG.DM}
\title{Find HVG with the DM approach}
\usage{
findHVG.DM(data, qMinThr = 0.9, qThr = 0.8, plot = FALSE)
}
\arguments{
\item{data}{Please provide the data without any transformation (e.g. log10).}

\item{qMinThr}{Quantile threshold to discard lowly expressed genes.E.g. if 
set to 0.9, it will take the data with median above the quantile 0.9 
and compute the DM. The ones below (bottom 90\%) will be discarded. Default: 0.9.}

\item{qThr}{Quantile threshold from which to consider a gene highly variable. E.g. If set 
to 0.8, it will take the most HVG genes falling on the top 20\%. Default: 0.8}

\item{plot}{Logical. Default: FALSE.}
}
\value{
vector of HVG.
}
\description{
It finds highly variable genes using the distance-to-median 
method proposed in Kolodziejczyk et al. Cell Stem Cell, 2015. I adapted the code from 
Ximena Ibarra-Soria.
}
\references{
Kolodziejczyk et al. Cell Stem Cell, 2015.
}
\author{
Blanca Pijuan-Sala.
}
