% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{genesHVG}
\alias{genesHVG}
\alias{genesHVG,RNAseq-method}
\title{Access genesHVG data}
\usage{
genesHVG(object)

\S4method{genesHVG}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
genesHVG.
}
\description{
Accesses \code{genesHVG} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
