% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{genesHVG<-}
\alias{genesHVG<-}
\alias{genesHVG<-,RNAseq,vector-method}
\title{Modify genesHVG data}
\usage{
genesHVG(object) <- value

\S4method{genesHVG}{RNAseq,vector}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new genesHVG data for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies genesHVG.
}
\description{
Modifies \code{genesHVG} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
