% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geneTable.R
\docType{data}
\name{geneTable}
\alias{geneTable}
\title{Ensembl geneIDs - geneNames}
\format{a \code{dataframe} with ENSMBL IDs and common gene names}
\source{
Ensembl bioMart (http://www.ensembl.org/biomart/).
}
\usage{
data(geneTable)
}
\description{
Relational table where column 1 are geneIDs and column2 associated 
geneNames.
}
\author{
Blanca Pijuan Sala
}
