% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meta_countsMatrix.R
\docType{data}
\name{meta}
\alias{meta}
\title{Example of metadata}
\source{
Blanca Pijuan-Sala (Gottgens Lab)
}
\usage{
data(meta)
}
\description{
Example of metadata
}
\author{
Blanca Pijuan-Sala
}
