% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{metadata}
\alias{metadata}
\alias{metadata,RNAseq-method}
\title{Access metadata}
\usage{
metadata(object)

\S4method{metadata}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
metadata
}
\description{
Accesses \code{metadata} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala
}
