% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{metadata<-}
\alias{metadata<-}
\alias{metadata<-,RNAseq,matrix-method}
\title{Modify metadata slot}
\usage{
metadata(object) <- value

\S4method{metadata}{RNAseq,matrix}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Matrix or dataframe containing new metadata for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies metadata.
}
\description{
Modifies \code{metadata} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala
}
