% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodRNAseq_normalise.R
\docType{methods}
\name{normalise}
\alias{normalise}
\alias{normalise,RNAseq-method}
\title{Normalise single cells from \code{RNAseq} object}
\usage{
normalise(object, cells = NULL, genes = NULL, genes.SizF = NULL,
  ercc.SizF = NULL)

\S4method{normalise}{RNAseq}(object, cells = NULL, genes = NULL,
  genes.SizF = NULL, ercc.SizF = NULL)
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{cells}{vector of cells to normalise. Default: NULL. This will first look at the
\code{cellsFailQC} slot and if empty, it will take all the cells in the 
\code{countsRaw} slot}

\item{genes}{vector of genes to normalise. Default: NULL (It will take all the genes of
the \code{countsRaw} slot).}

\item{genes.SizF}{Genes to use to compute genes size factors for each cell. Default: Vector 
of "genes" that are not ERCCs.}

\item{ercc.SizF}{ERCCs to use to compute ERCC size factors for each cell. Default: Vector 
of "erccs".}
}
\value{
List with normalised counts Matrix (countsNorm) and size factors for biological
genes (sizeFactorsGenes) and spike-ins (sizeFactorsSpikeIn). countsNorm should be stored in 
\code{object@countsNorm} slot and the size factors should be stored in 
\code{object@CellsSizeFac} slot as a list (with the two types of size factors).
}
\description{
It normalises single cells from countsRaw slot within the 
\code{RNAseq} object provided. It follows Brennecke et al., Nature Methods, 2013.
}
\author{
Blanca Pijuan Sala.
}
