% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{pca}
\alias{pca}
\alias{pca,RNAseq-method}
\title{Access pca data}
\usage{
pca(object)

\S4method{pca}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
pca.
}
\description{
Accesses \code{pca} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
