% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{pca<-}
\alias{pca<-}
\alias{pca<-,RNAseq-method}
\title{Modify pca data}
\usage{
pca(object) <- value

\S4method{pca}{RNAseq}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new pca data for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies pca.
}
\description{
Modifies \code{pca} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
