% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_quantileNorm.R
\name{quantileNorm}
\alias{quantileNorm}
\title{Quantile/Rank normalisation}
\usage{
quantileNorm(data, ties.rank = "min")
}
\arguments{
\item{data}{Dataframe or matrix containing the raw counts (genes x cells).}

\item{ties.rank}{It specifies how ties are treated. See \code{\link[base]{rank}} for
further details. Default: min.}
}
\description{
It normalises based on ranks.
}
\examples{

#test the function
data <- data.frame(one=c(5,2,3,4),
                   two=c(4,1,4,2),
                   three=c(3,4,6,8)
)
rownames(data) <- toupper(letters[1:4])
quantileNorm(data)
}
\references{
https://davetang.org/muse/2014/07/07/quantile-normalisation-in-r/
}
\author{
Blanca Pijuan-Sala.
}
