% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{class}
\name{RNAseq-class}
\alias{RNAseq-class}
\title{Class RNAseq}
\description{
Class \code{RNAseq} defines an RNAseq dataset.
}
\section{Slots}{

\describe{
\item{\code{countsRaw}}{matrix of raw counts (genes x cells). Recommended: Use gene IDs 
instead of gene Names.}

\item{\code{CountsFeat}}{matrix (features x cells) indicating the number of 
no feature reads (row 1), ambiguous reads (row 2), low quality reads (row 3),
unaligned reads (row 4) and others. The cellNames should be the same as in countsRaw.}

\item{\code{metadata}}{matrix of metadata of all cells and genes (cells x features).}

\item{\code{SpikeIn}}{If not NULL, contains logical vector with genes as names, 
where TRUE indicates that it's spike-in and FALSE indicates that it's a biological gene.}

\item{\code{countsNorm}}{matrix of normalised counts (genes x cells).Recommended: 
Use gene IDs instead of gene Names.}

\item{\code{CellsSizeFac}}{Contains size factors obtained in the normalisation.}

\item{\code{cellsFailQC}}{List of cells that have failed QC.}

\item{\code{genesHVG}}{Vector of highly variable genes.}

\item{\code{pca}}{If not NULL, contains PCA results.}

\item{\code{tsne}}{If not NULL, contains tSNE results.}

\item{\code{difmap}}{If not NULL, contains diffusion map results.}

\item{\code{clustering}}{If not NULL, contains information about potential clustering.}
}}

