% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodRNAseq_runCellQC.R
\docType{methods}
\name{runCellQC}
\alias{runCellQC}
\alias{runCellQC,RNAseq-method}
\title{Perform QC on cells from \code{RNAseq} object}
\usage{
runCellQC(object, annTable = geneTable, geneInput = "ID",
  GeneDetection = 2, minMapreads = 2e+05, maxMapmit = 0.1,
  maxMapSpike = 0.5, minGenesExpr = 4000, checkCountsFeat = TRUE,
  outputPlots = "./", plotting = "pdf", passCol = "grey70",
  failCol = "red", thresCol = "blue", lty = 3, lwd = 2)

\S4method{runCellQC}{RNAseq}(object, annTable = geneTable, geneInput = "ID",
  GeneDetection = 2, minMapreads = 2e+05, maxMapmit = 0.1,
  maxMapSpike = 0.5, minGenesExpr = 4000, checkCountsFeat = TRUE,
  outputPlots = "./", plotting = "pdf", passCol = "grey70",
  failCol = "red", thresCol = "blue", lty = 3, lwd = 2)
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{annTable}{Associative matrix or dataframe where column 1 reflects geneIDs
and column2 the associated geneName. Default: Mus musculus geneTable (??geneTable).}

\item{geneInput}{It states whether the genes in the counts table are
IDs ("ID") or associated gene names ("name"). Note that IDs will refer to column
1 of annTable and names to column 2 of annTable. In the case of "names": if the ID 
is not found, the gene will be discarded. Default: ID (recommended).}

\item{GeneDetection}{How many reads in a gene are needed to consider
a gene detected in a cell. Default: 2. Important for nDetectedGenes plot.}

\item{minMapreads}{Minimum number of reads mapping to nuclear genes.
Default: 200,000. Check the nNuclearGenes plot. Important: This plot is log-scaled.}

\item{maxMapmit}{Maximum percentage mapped mitochondrial reads. Default: 0.1.
Check the fMapped2Mit plot.}

\item{maxMapSpike}{Maximum percentage of mapped spikeins. Default: 0.5.If
no gene is spike-in, this condition will not be taken into account.
Check the fMapped2ERCCs plot.}

\item{minGenesExpr}{Minimum number of genes expressed in one cell at least. 
Default: 4000. Check nDetectedGenes plot.}

\item{checkCountsFeat}{Logical (TRUE/FALSE). Specifies whether the QC should take into account
the features of the mapping. Default = TRUE (recommended).}

\item{outputPlots}{state directory where you want to output the plots.
Default: Current directory.}

\item{plotting}{It states whether plots are generated or not and you should specify
the type. Options: "pdf", "tiff", "no". Default: "pdf". 
Default: TRUE}

\item{passCol}{Plotting parameter. Color for cells that have passed QC. 
Default: "grey70"}

\item{failCol}{Plotting parameter. Color for cells that have passed QC. 
Default: "red".}

\item{thresCol}{Plotting parameter. Color for threshold line. Default: "blue".}

\item{lty}{Plotting parameter. Line shape for threshold. For further information
see \code{\link[graphics]{par}}. Default: 3.}

\item{lwd}{Plotting parameter. Line width for threshold.For further information
see \code{\link[graphics]{par}}. Default: 2.}
}
\value{
vector of logical values where TRUE = fail and FALSE = pass
stored in \code{runCellQC} slot from \code{RNAseq} object.
}
\description{
It performs QC on cells from countsRaw slot within the 
\code{RNAseq} object provided. It follows Brennecke et al., Nature Methods, 2013.
}
\author{
Blanca Pijuan Sala.
}
