% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,RNAseq-method}
\title{Show details of \code{RNAseq} class}
\usage{
show(object)

\S4method{show}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
It gives general information on the RNAseq class.
}
\description{
Show details of \code{RNAseq} class.
}
\author{
Blanca Pijuan Sala.
}
