% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{tsne}
\alias{tsne}
\alias{tsne,RNAseq-method}
\title{Access tsne data}
\usage{
tsne(object)

\S4method{tsne}{RNAseq}(object)
}
\arguments{
\item{object}{\code{RNAseq} object.}
}
\value{
tsne.
}
\description{
Accesses \code{tsne} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
