% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RNAseq_accessorsANDreplace.R
\docType{methods}
\name{tsne<-}
\alias{tsne<-}
\alias{tsne<-,RNAseq-method}
\title{Modify tsne data}
\usage{
tsne(object) <- value

\S4method{tsne}{RNAseq}(object) <- value
}
\arguments{
\item{object}{\code{RNAseq} object.}

\item{value}{Object containing new tsne data for \code{RNAseq} object.}
}
\value{
It does not display any value but modifies tsne.
}
\description{
Modifies \code{tsne} slot in \code{RNAseq} class object.
}
\author{
Blanca Pijuan Sala.
}
